+++ src/term.h
@@ -74,7 +74,7 @@ void term_raw(void);
 void term_cooked(void);
 
 bool term_read_key(Key *key);
-char *term_read_paste(long *size);
+char *term_read_paste(size_t *size);
 void term_discard_paste(void);
 
 int term_get_size(int *w, int *h);
